airflow/requirements.txt
apache-airflow-providers-docker
boto3
